pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfw"
version = "0.1.0"
description = "Official code for the paper: Self-supervised Driver Distraction Detection for Imbalanced Datasets"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Suraj Bhardwaj"},
    {name = "David J. Lerch"},
    {name = "Manuel Martin"},
    {name = "Frederik Diederichs"},
    {name = "Rainer Stiefelhagen"},
]
keywords = ["deep learning", "computer vision", "class imbalance", "clustering", "driver distraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "Pillow>=9.4.0",
    "hdbscan>=0.8.33",
    "scikit-learn>=1.3.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tensorboard>=2.13.0",
    "tqdm>=4.65.0",
    "timm>=0.9.7",
    "transformers>=4.33.3",
    "opencv-python>=4.8.0",
    "requests>=2.28.0",
    "matplotlib>=3.7.0",
    "scipy>=1.11.0",
    "azure-ai-ml>=1.23.0",
    "azure-identity>=1.19.0",
    "mlflow==2.15.1",
    "azureml-mlflow==1.59.0.post1",
    "azure-storage-blob==12.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
bayesian = [
    "ray[default,tune]>=2.37.0",
    "hpbandster>=0.7.4",
    "ConfigSpace>=0.7.0",
]
azure = [
    "azure-ai-ml>=1.23.0",
    "azure-identity>=1.19.0",
    "mlflow==2.15.1",
    "azureml-mlflow==1.59.0.post1",
    "azure-storage-blob==12.19.0",
]
all = [
    "cfw[dev,bayesian,azure]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
markers = [
  "integration: marks integration tests",
  "slow: marks slow-running tests",
]

[tool.flake8]
max-line-length = 100
extend-ignore = "E203, W503"
exclude = ".git,__pycache__,build,dist,.eggs"
